refactor: 카테고리 드롭다운 로직 수정
