+++ src/components/feed/upload/CategorySelector/index.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/react';
 
-import CategorySelector from '@/components/feed/editor/CategorySelector';
+import CategorySelector from '@/components/feed/upload/CategorySelector';
 
 export default {
   component: CategorySelector,
